refactor: StringCalculator 패키지명 변경
